PaperbookToAudiobook/requirements.txt
google-cloud-vision>=0.42.0
google-cloud-texttospeech>=0.5.0
flask>=1.1.1
ffmpeg-python>=0.2.0
polyglot>=16.7.4
pycld2>=0.41
PyICU>=2.4.2






